+++ src/Smartstore.Core/Catalog/Attributes/Domain/ProductVariantAttributeCombination.cs
@@ -36,10 +36,12 @@ namespace Smartstore.Core.Catalog.Attributes
     [Index(nameof(Gtin), Name = "IX_Gtin")]
     [Index(nameof(ManufacturerPartNumber), Name = "IX_ManufacturerPartNumber")]
     [Index(nameof(StockQuantity), nameof(AllowOutOfStockOrders), Name = "IX_StockQuantity_AllowOutOfStockOrders")]
+    [Index(nameof(HashCode), Name = "IX_HashCode")]
     public partial class ProductVariantAttributeCombination : BaseEntity, IAttributeAware
     {
         private ProductVariantAttributeSelection _attributeSelection;
         private string _rawAttributes;
+        private int? _hashCode;
 
         /// <summary>
         /// Gets or sets the product identifier.
@@ -157,8 +159,16 @@ namespace Smartstore.Core.Catalog.Attributes
             get => _rawAttributes;
             set
             {
+                var updateHash = !_rawAttributes.EqualsNoCase(value);
+
                 _rawAttributes = value;
                 _attributeSelection = null;
+
+                // Avoids the caller having to call GetAttributesHashCode itself on instantiation.
+                if (updateHash)
+                {
+                    _hashCode = AttributeSelection.GetHashCode();
+                }
             }
         }
 
@@ -175,6 +185,20 @@ namespace Smartstore.Core.Catalog.Attributes
         /// Gets or sets a value indicating whether to allow orders when out of stock.
         /// </summary>
         public bool AllowOutOfStockOrders { get; set; }
+
+        /// <summary>
+        /// Gets or sets the attributes hash code. This is generally equal to <see cref="AttributeSelection.GetHashCode"/>.
+        /// </summary>
+        [Required]
+        public int HashCode
+        {
+            get => GetAttributesHashCode();
+            // Setter for EF.
+            set => _hashCode = value;
+        }
+
+        public int GetAttributesHashCode()
+            => _hashCode ??= AttributeSelection.GetHashCode();
 
         /// <summary>
         /// Gets the assigned media file identifiers.